Relocate Code class inside MessagePack again since it is used for Packer/Unpacker, which do not need to depend on MessageFormat enum
